+++ opencascade-hs/cpp/hs_BRepAdaptor_Curve.cpp
@@ -13,3 +13,11 @@ void hs_delete_BRepAdaptor_Curve(BRepAdaptor_Curve * curve){
 GeomAbs_CurveType hs_BRepAdaptor_Curve_curveType(BRepAdaptor_Curve* curve){
     return curve->GetType();   
 }
+
+Handle(Geom_BezierCurve) * hs_BRepAdaptor_Curve_bezier(BRepAdaptor_Curve * curve){
+    return new opencascade::handle(curve->Bezier());
+}
+
+Handle(Geom_BSplineCurve) * hs_BRepAdaptor_Curve_bspline(BRepAdaptor_Curve * curve){
+    return new opencascade::handle(curve->BSpline());
+}