+++ opencascade-hs/CHANGELOG.md
@@ -15,7 +15,8 @@ and this project adheres to the
 - Add `OpenCascade.BRepAdaptor.Curve`
 - Add `OpenCascade.GeomAbs.CurveType`
 - Add `OpenCascade.GeomConvert.BSplineCurveToBezierCurve`
-- Add `OpenCascade.Geom.BezierCurve` methods `pole` and `nbPoles`
+- Add `OpenCascade.Geom.BSplineCurve`
+- Add `OpenCascade.Geom.BezierCurve` methods `pole`, `nbPoles` and `isRational`
 
 ## 0.4.0.0